+++ lib/msal-browser/src/protocol/Authorize.ts
@@ -43,7 +43,6 @@ import { PlatformAuthInteractionClient } from "../interaction_client/PlatformAut
 import { EventHandler } from "../event/EventHandler.js";
 import { decryptEarResponse } from "../crypto/BrowserCrypto.js";
 import { IPlatformAuthHandler } from "../broker/nativeBroker/IPlatformAuthHandler.js";
-import { createForm } from "../utils/BrowserUtils.js";
 
 /**
  * Parsed representation of the clientdata response parameter from the /authorize endpoint.
@@ -429,6 +428,38 @@ export async function getCodeForm(
         performanceClient
     );
     return createForm(frame, action, fields);
+}
+
+/**
+ * Creates a POST `<form>` in the provided document with the given auth
+ * parameters as hidden inputs, targeting `action`. Shared by the popup auth
+ * flows (`getEARForm` / `getCodeForm`) and the popup-relay page
+ * (`runPopupRelay`) so the form is built one way only.
+ *
+ * @param frame - document to create the form in
+ * @param action - form action (the /authorize URL)
+ * @param fields - POST-body fields
+ * @internal
+ */
+export function createForm(
+    frame: Document,
+    action: string,
+    fields: Record<string, string>
+): HTMLFormElement {
+    const form = frame.createElement("form");
+    form.method = "post";
+    form.action = action;
+
+    Object.keys(fields).forEach((name) => {
+        const input = frame.createElement("input");
+        input.hidden = true;
+        input.name = name;
+        input.value = fields[name];
+        form.appendChild(input);
+    });
+
+    frame.body.appendChild(form);
+    return form;
 }
 
 /**
